+++ converters/fromZigbee.js
@@ -4811,100 +4811,6 @@ const converters = {
             }
         },
     },
-    x5h_thermostat: {
-        cluster: 'manuSpecificTuya',
-        type: ['commandDataResponse', 'commandDataReport'],
-        convert: (model, msg, publish, options, meta) => {
-            const dpValue = tuya.firstDpValue(msg, meta, 'x5h_thermostat');
-            const dp = dpValue.dp;
-            const value = tuya.getDataValue(dpValue);
-
-            switch (dp) {
-            case tuya.dataPoints.x5hState: {
-                return {system_mode: value ? 'heat' : 'off'};
-            }
-            case tuya.dataPoints.x5hWorkingStatus: {
-                return {running_state: value ? 'heat' : 'idle'};
-            }
-            case tuya.dataPoints.x5hSound: {
-                return {sound: value ? 'ON' : 'OFF'};
-            }
-            case tuya.dataPoints.x5hFrostProtection: {
-                return {frost_protection: value ? 'ON' : 'OFF'};
-            }
-            case tuya.dataPoints.x5hWorkingDaySetting: {
-                return {week: tuya.thermostatWeekFormat[value]};
-            }
-            case tuya.dataPoints.x5hFactoryReset: {
-                if (value) {
-                    clearTimeout(globalStore.getValue(msg.endpoint, 'factoryResetTimer'));
-                    const timer = setTimeout(() => publish({factory_reset: 'OFF'}), 60 * 1000);
-                    globalStore.putValue(msg.endpoint, 'factoryResetTimer', timer);
-                    meta.logger.info('The thermostat is resetting now. It will be available in 1 minute.');
-                }
-
-                return {factory_reset: value ? 'ON' : 'OFF'};
-            }
-            case tuya.dataPoints.x5hTempDiff: {
-                return {temp_diff: parseFloat((value / 10).toFixed(1))};
-            }
-            case tuya.dataPoints.x5hProtectionTempLimit: {
-                return {protection_temp_limit: value};
-            }
-            case tuya.dataPoints.x5hBackplaneBrightness: {
-                const lookup = {0: 'off', 1: 'low', 2: 'medium', 3: 'high'};
-
-                if (value >= 0 && value <= 3) {
-                    globalStore.putValue(msg.endpoint, 'brightnessState', value);
-                    return {brightness_state: lookup[value]};
-                }
-
-                // Sometimes, for example on thermostat restart, it sends message like:
-                // {"dpValues":[{"data":{"data":[90],"type":"Buffer"},"datatype":4,"dp":104}
-                // It doesn't represent any brightness value and brightness remains the previous value
-                const lastValue = globalStore.getValue(msg.endpoint, 'brightnessState') || 1;
-                return {brightness_state: lookup[lastValue]};
-            }
-            case tuya.dataPoints.x5hWeeklyProcedure: {
-                return;
-            }
-            case tuya.dataPoints.x5hChildLock: {
-                return {child_lock: value ? 'LOCK' : 'UNLOCK'};
-            }
-            case tuya.dataPoints.x5hSetTemp: {
-                return {current_heating_setpoint: parseFloat((value / 10).toFixed(1))};
-            }
-            case tuya.dataPoints.x5hSetTempCeiling: {
-                // It overwrites heating setpoint
-                return {upper_temp: value / 10};
-            }
-            case tuya.dataPoints.x5hCurrentTemp: {
-                const temperature = value & (1 << 15) ? value - (1 << 16) + 1 : value;
-                return {local_temperature: parseFloat((temperature / 10).toFixed(1))};
-            }
-            case tuya.dataPoints.x5hTempCorrection: {
-                let temperature = value;
-
-                if (temperature > 4000) {
-                    temperature = temperature - 4096;
-                }
-
-                return {local_temperature_calibration: parseFloat((temperature / 10).toFixed(1))};
-            }
-            case tuya.dataPoints.x5hMode: {
-                const lookup = {0: 'manual', 1: 'program'};
-                return {preset: lookup[value]};
-            }
-            case tuya.dataPoints.x5hSensorSelection: {
-                const lookup = {0: 'internal', 1: 'external', 2: 'both'};
-                return {sensor: lookup[value]};
-            }
-            default: {
-                meta.logger.warn(`fromZigbee:x5h_thermostat: Unrecognized DP #${dp} with data ${JSON.stringify(dpValue)}`);
-            }
-            }
-        },
-    },
     tuya_dimmer: {
         cluster: 'manuSpecificTuya',
         type: ['commandDataResponse', 'commandDataReport'],
