+++ converters/toZigbee.js
@@ -4957,102 +4957,6 @@ const converters = {
             return {state: {week: value}};
         },
     },
-    x5h_thermostat: {
-        key: ['system_mode', 'current_heating_setpoint', 'sensor', 'brightness_state', 'sound', 'frost_protection', 'week', 'factory_reset',
-            'local_temperature_calibration', 'protection_temp_limit', 'temp_diff', 'upper_temp', 'preset', 'child_lock'],
-        convertSet: async (entity, key, value, meta) => {
-            switch (key) {
-            case 'system_mode':
-                await tuya.sendDataPointBool(entity, tuya.dataPoints.x5hState, value === 'heat');
-                break;
-            case 'preset': {
-                value = value.toLowerCase();
-                const lookup = {manual: 0, program: 1};
-                utils.validateValue(value, Object.keys(lookup));
-                value = lookup[value];
-                await tuya.sendDataPointEnum(entity, tuya.dataPoints.x5hMode, value);
-                break;
-            }
-            case 'upper_temp':
-                if (value >= 35 && value <= 95) {
-                    await tuya.sendDataPointValue(entity, tuya.dataPoints.x5hSetTempCeiling, value);
-                } else {
-                    throw new Error('Supported values are in range [35, 95]');
-                }
-                break;
-            case 'temp_diff':
-                if (value >= 1 && value <= 9.5) {
-                    value = Math.round(value * 10);
-                    await tuya.sendDataPointValue(entity, tuya.dataPoints.x5hTempDiff, value);
-                } else {
-                    throw new Error('Supported values are in range [1, 9.5]');
-                }
-                break;
-            case 'protection_temp_limit':
-                if (value >= 5 && value <= 60) {
-                    await tuya.sendDataPointValue(entity, tuya.dataPoints.x5hProtectionTempLimit, value);
-                } else {
-                    throw new Error('Supported values are in range [5, 60]');
-                }
-                break;
-            case 'local_temperature_calibration':
-                if (value >= -9.9 && value <= 9.9) {
-                    value = Math.round(value * 10);
-
-                    if (value < 0) {
-                        value = 4096 + value;
-                    }
-
-                    await tuya.sendDataPointValue(entity, tuya.dataPoints.x5hTempCorrection, value);
-                } else {
-                    throw new Error('Supported values are in range [-9.9, 9.9]');
-                }
-                break;
-            case 'factory_reset':
-                await tuya.sendDataPointBool(entity, tuya.dataPoints.x5hFactoryReset, value === 'ON');
-                break;
-            case 'week':
-                await tuya.sendDataPointEnum(entity, tuya.dataPoints.x5hWorkingDaySetting,
-                    utils.getKey(tuya.thermostatWeekFormat, value, value, Number));
-                break;
-            case 'frost_protection':
-                await tuya.sendDataPointBool(entity, tuya.dataPoints.x5hFrostProtection, value === 'ON');
-                break;
-            case 'sound':
-                await tuya.sendDataPointBool(entity, tuya.dataPoints.x5hSound, value === 'ON');
-                break;
-            case 'brightness_state': {
-                value = value.toLowerCase();
-                const lookup = {off: 0, low: 1, medium: 2, high: 3};
-                utils.validateValue(value, Object.keys(lookup));
-                value = lookup[value];
-                await tuya.sendDataPointEnum(entity, tuya.dataPoints.x5hBackplaneBrightness, value);
-                break;
-            }
-            case 'sensor': {
-                value = value.toLowerCase();
-                const lookup = {'internal': 0, 'external': 1, 'both': 2};
-                utils.validateValue(value, Object.keys(lookup));
-                value = lookup[value];
-                await tuya.sendDataPointEnum(entity, tuya.dataPoints.x5hSensorSelection, value);
-                break;
-            }
-            case 'current_heating_setpoint':
-                if (value >= 5 && value <= 60) {
-                    value = Math.round(value * 10);
-                    await tuya.sendDataPointValue(entity, tuya.dataPoints.x5hSetTemp, value);
-                } else {
-                    throw new Error(`Unsupported value: ${value}`);
-                }
-                break;
-            case 'child_lock':
-                await tuya.sendDataPointBool(entity, tuya.dataPoints.x5hChildLock, value === 'LOCK');
-                break;
-            default:
-                break;
-            }
-        },
-    },
     tuya_cover_control: {
         key: ['state', 'position'],
         options: [exposes.options.invert_cover()],
